create yaml config, cleanup training script
